refactor(fuse): extract chart to a separate component
